pyproject.toml
[tool.poetry]
name = "todo-everything-django"
version = "0.1.0"
description = ""
authors = ["Allan Almazan <allan.almazan@gmail.com>"]

[tool.isort]
profile = "black"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.6"
django-environ = "^0.11.2"
django-model-utils = "^4.0.0"
django-crispy-forms = "^2.0"
psycopg2 = "^2.9.9"
djangorestframework = "^3.14.0"
django-filter = "^23.3"
djangorestframework-simplejwt = {extras = ["crypto"], version = "^5.3.0"}
celery = "^5.3.4"

[tool.poetry.group.dev.dependencies]
black = "^23.9.1"
pytest-factoryboy = "^2.0.3"
hypothesis = "^6.87.1"
ipython = "^8.16.1"
pytest-cov = "^4.1.0"
pytest-django = "^4.5.2"
isort = "^5.12.0"
flake8 = "^6.1.0"
doc8 = "^1.1.1"
autopep8 = "^2.0.4"
pytest = "^7.4.2"
pytest-socket = "^0.6.0"
model-bakery = "^1.15.0"
invoke = "^2.2.0"
pre-commit = "^3.4.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
